feature (resource): parse interval value

+++ src/models/v2/resources/values/read-decimal-value.ts
@@ -7,5 +7,5 @@ import {DecimalConverter} from '../../CustomConverters';
 export class ReadDecimalValue extends ReadValue {
 
     @JsonProperty(Constants.DecimalValueAsDecimal, DecimalConverter)
-    int: number = 0;
+    decimal: number = 0;
 }
